DELIA-22919: Enabling SetBluetoothDiscoverable timeout to set to infinite
Reason for change: Providing option to set Discoverable timeout to no timeout
Test Procedure: The default timeout is 180 secs, to set it to no timeout use BtMgrTest set Discoverable option and set timeout 0, to set default timeout input -1.
Risks: medium

+++ include/btmgr.h
@@ -271,7 +271,7 @@ BTRMGR_Result_t BTRMGR_GetAdapterName(unsigned char index_of_adapter, char* pNam
 BTRMGR_Result_t BTRMGR_SetAdapterPowerStatus(unsigned char index_of_adapter, unsigned char power_status);
 BTRMGR_Result_t BTRMGR_GetAdapterPowerStatus(unsigned char index_of_adapter, unsigned char *pPowerStatus);
 
-BTRMGR_Result_t BTRMGR_SetAdapterDiscoverable(unsigned char index_of_adapter, unsigned char discoverable, unsigned short timeout);
+BTRMGR_Result_t BTRMGR_SetAdapterDiscoverable(unsigned char index_of_adapter, unsigned char discoverable, int timeout);
 BTRMGR_Result_t BTRMGR_IsAdapterDiscoverable(unsigned char index_of_adapter, unsigned char *pDiscoverable);
 
 BTRMGR_Result_t BTRMGR_StartDeviceDiscovery(unsigned char index_of_adapter);
